# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Anderson Personal Trainer website
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: appointments
    description: Appointment scheduling
  - name: testimonials
    description: Student testimonials and before/after results
  - name: blog
    description: Blog articles
  - name: contacts
    description: Contact form submissions
  - name: admin
    description: Admin dashboard and management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List all appointments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, cancelled]
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Schedule a new evaluation appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          description: Validation error

  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /appointments/stats:
    get:
      operationId: getAppointmentStats
      tags: [appointments]
      summary: Get appointment statistics for admin dashboard
      responses:
        "200":
          description: Appointment stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentStats"

  /testimonials:
    get:
      operationId: listTestimonials
      tags: [testimonials]
      summary: List approved testimonials
      parameters:
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of testimonials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"
    post:
      operationId: createTestimonial
      tags: [testimonials]
      summary: Create a new testimonial (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialInput"
      responses:
        "201":
          description: Testimonial created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"

  /testimonials/{id}:
    patch:
      operationId: updateTestimonial
      tags: [testimonials]
      summary: Update a testimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
    delete:
      operationId: deleteTestimonial
      tags: [testimonials]
      summary: Delete a testimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /blog:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List published blog posts
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: published
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create a new blog post (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "201":
          description: Blog post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"

  /blog/{id}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get blog post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blog post found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found
    patch:
      operationId: updateBlogPost
      tags: [blog]
      summary: Update a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    delete:
      operationId: deleteBlogPost
      tags: [blog]
      summary: Delete a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List contact form submissions (admin)
      parameters:
        - name: read
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Submit a contact form message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Contact submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"

  /contacts/{id}:
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Mark contact as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"

  /admin/dashboard:
    get:
      operationId: getDashboardStats
      tags: [admin]
      summary: Get admin dashboard overview metrics
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Appointment:
      type: object
      required: [id, name, phone, email, goal, desiredDate, desiredTime, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        goal:
          type: string
          enum: [emagrecimento, hipertrofia, condicionamento, iniciantes, outro]
        desiredDate:
          type: string
        desiredTime:
          type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        createdAt:
          type: string

    AppointmentInput:
      type: object
      required: [name, phone, email, goal, desiredDate, desiredTime]
      properties:
        name:
          type: string
          minLength: 2
        phone:
          type: string
          minLength: 8
        email:
          type: string
        goal:
          type: string
          enum: [emagrecimento, hipertrofia, condicionamento, iniciantes, outro]
        desiredDate:
          type: string
        desiredTime:
          type: string
        notes:
          type: string

    AppointmentUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        notes:
          type: string

    AppointmentStats:
      type: object
      required: [total, pending, confirmed, cancelled, thisMonth]
      properties:
        total:
          type: integer
        pending:
          type: integer
        confirmed:
          type: integer
        cancelled:
          type: integer
        thisMonth:
          type: integer

    Testimonial:
      type: object
      required: [id, studentName, goalType, content, weightLost, duration, featured, approved, createdAt]
      properties:
        id:
          type: integer
        studentName:
          type: string
        goalType:
          type: string
          enum: [emagrecimento, hipertrofia, condicionamento, iniciantes]
        content:
          type: string
        weightLost:
          type: ["number", "null"]
        duration:
          type: string
        beforeImageUrl:
          type: ["string", "null"]
        afterImageUrl:
          type: ["string", "null"]
        featured:
          type: boolean
        approved:
          type: boolean
        createdAt:
          type: string

    TestimonialInput:
      type: object
      required: [studentName, goalType, content, duration]
      properties:
        studentName:
          type: string
        goalType:
          type: string
          enum: [emagrecimento, hipertrofia, condicionamento, iniciantes]
        content:
          type: string
        weightLost:
          type: number
        duration:
          type: string
        beforeImageUrl:
          type: string
        afterImageUrl:
          type: string
        featured:
          type: boolean
        approved:
          type: boolean

    TestimonialUpdate:
      type: object
      properties:
        studentName:
          type: string
        goalType:
          type: string
          enum: [emagrecimento, hipertrofia, condicionamento, iniciantes]
        content:
          type: string
        weightLost:
          type: number
        duration:
          type: string
        beforeImageUrl:
          type: string
        afterImageUrl:
          type: string
        featured:
          type: boolean
        approved:
          type: boolean

    BlogPost:
      type: object
      required: [id, title, slug, category, excerpt, content, published, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        category:
          type: string
          enum: [treino, alimentacao, emagrecimento, hipertrofia, saude]
        excerpt:
          type: string
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        published:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string

    BlogPostInput:
      type: object
      required: [title, slug, category, excerpt, content]
      properties:
        title:
          type: string
          minLength: 3
        slug:
          type: string
        category:
          type: string
          enum: [treino, alimentacao, emagrecimento, hipertrofia, saude]
        excerpt:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        published:
          type: boolean

    BlogPostUpdate:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        category:
          type: string
          enum: [treino, alimentacao, emagrecimento, hipertrofia, saude]
        excerpt:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        published:
          type: boolean

    Contact:
      type: object
      required: [id, name, email, message, read, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        message:
          type: string
        read:
          type: boolean
        createdAt:
          type: string

    ContactInput:
      type: object
      required: [name, email, message]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
        phone:
          type: string
        subject:
          type: string
        message:
          type: string
          minLength: 10

    ContactUpdate:
      type: object
      properties:
        read:
          type: boolean

    DashboardStats:
      type: object
      required: [totalAppointments, pendingAppointments, totalTestimonials, totalBlogPosts, totalContacts, unreadContacts, appointmentsThisMonth, contactsThisMonth]
      properties:
        totalAppointments:
          type: integer
        pendingAppointments:
          type: integer
        confirmedAppointments:
          type: integer
        totalTestimonials:
          type: integer
        totalBlogPosts:
          type: integer
        publishedBlogPosts:
          type: integer
        totalContacts:
          type: integer
        unreadContacts:
          type: integer
        appointmentsThisMonth:
          type: integer
        contactsThisMonth:
          type: integer