# lib/api-zod/src/generated/types/appointmentStats.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Anderson Personal Trainer website
 * OpenAPI spec version: 0.1.0
 */

export interface AppointmentStats {
  total: number;
  pending: number;
  confirmed: number;
  cancelled: number;
  thisMonth: number;
}