# lib/api-zod/src/generated/types/appointmentUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Anderson Personal Trainer website
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentUpdateStatus } from './appointmentUpdateStatus';

export interface AppointmentUpdate {
  status?: AppointmentUpdateStatus;
  notes?: string;
}