# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Anderson Personal Trainer website
 * OpenAPI spec version: 0.1.0
 */
import type { BlogPostCategory } from './blogPostCategory';

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  category: BlogPostCategory;
  excerpt: string;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  published: boolean;
  createdAt: string;
  updatedAt?: string;
}