# lib/api-zod/src/generated/types/contact.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Anderson Personal Trainer website
 * OpenAPI spec version: 0.1.0
 */

export interface Contact {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  subject?: string | null;
  message: string;
  read: boolean;
  createdAt: string;
}