# lib/api-zod/src/generated/types/testimonial.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Anderson Personal Trainer website
 * OpenAPI spec version: 0.1.0
 */
import type { TestimonialGoalType } from './testimonialGoalType';

export interface Testimonial {
  id: number;
  studentName: string;
  goalType: TestimonialGoalType;
  content: string;
  /** @nullable */
  weightLost: number | null;
  duration: string;
  /** @nullable */
  beforeImageUrl?: string | null;
  /** @nullable */
  afterImageUrl?: string | null;
  featured: boolean;
  approved: boolean;
  createdAt: string;
}